Finished project...?

+++ src/modules/daybook/views/EntryView.vue
@@ -8,6 +8,14 @@
       </div>
 
       <div>
+        <input
+          type="file"
+          @change="onSelectedImage"
+          ref="imageSelector"
+          v-show="false"
+          accept="imaga/png, image/jpeg, image/png"
+        />
+
         <button
           v-if="entry.id"
           class="btn btn-danger mx-2"
@@ -16,7 +24,7 @@
           Borrar
           <i class="fa fa-trash-alt"></i>
         </button>
-        <button class="btn btn-primary">
+        <button class="btn btn-primary" @click="onSelectImage">
           Subir foto
           <i class="fa fa-upload"></i>
         </button>
@@ -30,7 +38,14 @@
     </div>
 
     <img
-      src="https://s.yimg.com/ny/api/res/1.2/NO32AnKeYxKJlrw5RKk2eg--/YXBwaWQ9aGlnaGxhbmRlcjt3PTk2MDtoPTU0MDtjZj13ZWJw/https://s.yimg.com/os/creatr-uploaded-images/2022-05/c86d1780-d79e-11ec-97ff-f8cb09fc00e6"
+      v-if="entry.picture && !localImage"
+      :src="entry.picture"
+      alt="entry-pic"
+      class="img-thumbnail"
+    />
+    <img
+      v-if="localImage"
+      :src="localImage"
       alt="entry-pic"
       class="img-thumbnail"
     />
@@ -42,7 +57,9 @@
 <script>
 import { defineAsyncComponent } from "vue";
 import { mapGetters, mapActions } from "vuex";
+import Swal from "sweetalert2";
 import getDayMonthYear from "../helpers/getDayMonthYear";
+import uploadImage from "../helpers/uploadImage";
 
 export default {
   props: {
@@ -55,6 +72,8 @@ export default {
   data() {
     return {
       entry: null,
+      localImage: null,
+      file: null,
     };
   },
 
@@ -83,6 +102,7 @@ export default {
 
     loadEntry() {
       let entry;
+      this.localImage = null;
       if (this.id == "new") {
         entry = {
           text: "",
@@ -97,17 +117,63 @@ export default {
     },
 
     async saveEntry() {
+      new Swal({
+        title: "Espere por favor",
+        allowOutsideClick: false,
+      });
+      Swal.showLoading();
+
+      const picture = await uploadImage(this.file);
+      this.entry.picture = picture;
+
       if (this.entry.id) {
         await this.updateEntry(this.entry);
       } else {
         const id = await this.createEntry(this.entry);
         this.$router.push({ name: "entry", params: { id } });
       }
+
+      this.file = null;
+      Swal.fire("Guardado", "Entrada registrada con éxito", "success");
     },
 
     async onDeleteEntry() {
-      await this.deleteEntry(this.entry.id);
-      this.$router.push({ name: "no-entry" });
+      const { isConfirmed } = await Swal.fire({
+        title: "¿Está seguro",
+        text: "Una vez borrado, no se puede recuperar",
+        showDenyButton: true,
+        confirmButtonText: "Si, estoy seguro",
+      });
+      if (isConfirmed) {
+        new Swal({
+          title: "Espere por favor",
+          allowOutsideClick: false,
+        });
+        Swal.showLoading();
+        await this.deleteEntry(this.entry.id);
+        this.$router.push({ name: "no-entry" });
+
+        Swal.fire("Eliminado", "", "success");
+      }
+    },
+
+    onSelectedImage(event) {
+      const file = event.target.files[0];
+      if (!file) {
+        this.localImage = null;
+        this.file = null;
+        return;
+      }
+
+      this.file = file;
+
+      const fr = new FileReader();
+      fr.onload = () => (this.localImage = fr.result);
+      fr.readAsDataURL(file);
+    },
+
+    onSelectImage() {
+      this.$refs["imageSelector"].click();
     },
   },
